backend/requirements.txt
fastapi
uvicorn
pydantic
chromadb
google-generativeai
pypdf
python-multipart
python-dotenv
slowapi
fpdf2
